backend/pyproject.toml
[tool.poetry]
name = "catcord"
version = "0.1.0"
description = ""
authors = ["Cat Developer Group"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.67.0"
uvicorn = "^0.14.0"
asyncpg = "^0.23.0"
gunicorn = "^20.1.0"
starlette = "^0.14.2"
aiofiles = "^0.7.0"
taskipy = "^1.8.1" # This is a dep because the dockerfile uses taskipy and installs without dev deps
loguru = "^0.5.3"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.6"}
aerich = "^0.5.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^21.7b0"
flake8 = "^3.9.2"
isort = "^5.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "poetry run gunicorn src.main:app -w 4 -k uvicorn.workers.UvicornWorker --bind 0.0.0.0"
lint = "poetry run flake8"
format = "poetry run black . && poetry run isort ."
test = "poetry run pytest"